refactor(tui): extract semantic text rendering component
